requirements.txt

pandas 
dvc
mlflow==2.2.2
notebook
numpy
matplotlib
seaborn
scikit-learn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
gdown
Pillow
setuptools
